+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 -----------------------
 
 - fix test suite with --no-test-cache run
+- disable warnings for pytest_localserver's smtp support
 
 0.8.3
 -----------------------